apps/wiki-chat/pyproject.toml
[project]
name = "liz-wiki-chat"
version = "0.1.0"
description = "Conversational chat over the Liz project wiki"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
